Implement reset functionality for process registry
- Added a new `Reset` RPC method to the GoProc service for clearing the process registry and resetting ID counters.
- Introduced `ResetRequest` and `ResetResponse` message types in the protocol buffer definitions.
- Implemented the reset logic in the daemon service, ensuring proper state management.
- Enhanced the command-line interface with a `reset` command, requiring user confirmation for destructive actions.
- Updated existing commands to handle `uint64` IDs for consistency across the application.

+++ cmd/goproc/list.go
@@ -74,7 +74,7 @@ var cmdList = &cobra.Command{
 			if id <= 0 {
 				return fmt.Errorf("invalid id filter: %d", id)
 			}
-			req.Ids = append(req.Ids, uint32(id))
+			req.Ids = append(req.Ids, uint64(id))
 		}
 
 		resp, err := client.List(ctx, req)
